Expose the http status code of the response on the error object
Currently, authorization problems with the API return an
`invalid_request_error`.  This makes it impossible to distinguish
a problem with your keys from an invalid request body without
parsing the error message or reading the HTTP status code.
Since we don't commit to maintaining consistent error messages,
it would be helpful to have the status code accessible on the
error objects we return.

Alternatively, we could diverge somewhat from the API and return
 `Type: Unauthorized` in this case.