Deprecate `Zip` in favor of `PostalCode`
Deprecates an address' the `Zip` property in favor of `PostalCode`, but
for now maintains `Zip` for backward compatibility.

The impetus for this is that the bindings are already inconsistent with
the API in that this property is serialized under a different name, but
as described in #155 are also self-inconsistent in that the property is
just called `PostalCode` elsewhere.