Allow `Account` to be specified for `ListParams`
As discussed in #196, we cannot currently send an account header when
using list endpoints. I believe that this was a known limitation of the
original implementation that was never addressed because of subtle
differences in the way `Params` and `ListParams` are used (the former is
serialized around an `http.Request` object while the latter is just used
to build a set of parameter values).

This patch adds an `Account` field to `ListParams` and suggests that
it's injected in list callbacks using a new `ToParams` function on
`ListParams`. This implementation is not complete, and just demonstrates
its use on a single endpoint. The new field will have to be added to all
list callbacks to be fully workable.