Fix collection of miscellaneous linting errors
Many of these are just additions of documentation, but many are also
legitimate fixes (largely comments that have drifted out of date with
the naming of the construct that they're commenting).

Unfortunately, this job isn't even close to being done. It fixes many
problems in the top level package (except around Relay resources, which
aren't well documented), but all the client subpackages are still in bad
shape.