Makes routing_number only be sent in POST if set
When tokenizing a bank account, the routing number is only required
for US bank accounts. If using an IBAN for `account_number`, then
this field is optional. Prior to this commit, `routing_number` was
always being sent, even as a an empty string which was causing the
following error to be sent back from the Stripe API:

{
  "type":"invalid_request_error",
  "message":"Invalid bic",
  "param":"bank_account",
  "request_id":"req_XXXYYYZZZ",
  "status":400
}

Signed-off-by: Zach Wick <zwick@stripe.com>