Add a StableForm type that can support ordered duplicates
As described in #251, using `url.Values` to contain form values is
problematic for the Stripe API because it depends on "Rack style" form
encoding that requires duplicated parameters to occur in a certain order
for arrays to be correctly encoded.

This patch adds a new type called `StableForm` that encodes parameters
in the order that they were added so that arrays can be properly
supported. It also adds a `Form` interface that's shared between
`StableForm` and `url.Values` so that we can demonstrate this concept
for now without updating every use of `url.Values` everywhere. For
consistency, before the final merge we should probably move to
`StableForm` everywhere and remove the interface (which will produce a
messy diff).

Fixes #251.