Allow bank account operations to keyed to customer
Currently, we've got bank account operations tied specifically to
accounts, when they actually should be keyable to a customer as well
(for some reason delete was implemented as expected, but this goes for
list/retrieve/create). This meant that some of the code that we had in
our documentation wasn't valid:

    card.List(&stripe.BankAccountListParams{Customer: {CUSTOMER_ID}})

This patch allows the library to switch off of a customer key as well as
an account key.

Fixes #312.