Add `NoDiscountable` to `InvoiceItemsParams`
Currently, `Discountable` cannot be explicitly set to `false`, which is
a problem because for certain types of invoice items it will be
implicitly set to true on the server side.

This patch adds `NoDiscountable` which allows an explicit `false`. This
is modeled after current SDK convention (for exmpale `NoClosed` as part
of `InvoiceParams`).

Fixes #329.