Start test suite refactor + power tests with external API stub
In the style of https://github.com/stripe/stripe-ruby/pull/553 starts to
move toward an external API stub [1] executable to power the test suite.
This will allow us to start turning more tests back on in CI (currently
the vast majority of them don't run).

This is an experimental prototype for demonstrative purposes and should not be merged.

[1] https://github.com/brandur/stripestub