Alphabetize the world
Alphabetizes all API parameter and resource structs. I'm not going to
pretend like this was a great use of time, but at least it will allow us
to have some kind of convention to point to so that we can help control
consistency in the project.

This is technically breaking because you can initialize structs with
positional order in Go (e.g. `MyStruct{1, 2}`), but I doubt it'll break
almost anyone because (1) most people won't use the positional
initialization for these sorts of big structs, and (2) even if they
were, type mismatches in the new order will let the compiler tell them
about the problem. That said, there are a few potential danger spots
that I noticed like `PackageDimemsions` for Relay which is four fields
with all exactly the same type.

Let's put a note in the changelog but ship it as a major version. I
haven't released #446 yet, so I'd like to group it in with that.

Fixes part of #449.