Remove Displayer and Display
At some point we added a `Displayer` interface that worked with certain
payment sources to display a short human readable string.

This hasn't turned out to be maintained very well, and is a convention
that's so localized to one small part of the library that it's not very
useful in practice. Here we drop `Displayer` and its associated
implementations.

Fixes part of #449.