Build a type encoder caching system for the `form` package
Here we optimize the `AppendTo` function of the form package and its
associated helpers. As #446 came in, there was some concerns expressed
around its performance. This implementation might not be as quick as a
bunch of `if` statements, but it's pretty fast.

Before:

    $ make bench
    go test -bench . -run "Benchmark" ./form
    goos: darwin
    goarch: amd64
    pkg: github.com/stripe/stripe-go/form
    BenchmarkAppendTo-4       500000              2767 ns/op
    PASS
    ok      github.com/stripe/stripe-go/form        1.480s

After:

    $ make bench
    go test -bench . -run "Benchmark" ./form
    goos: darwin
    goarch: amd64
    pkg: github.com/stripe/stripe-go/form
    BenchmarkAppendTo-4      2000000               922 ns/op
    PASS
    ok      github.com/stripe/stripe-go/form        2.867s

Which shows roughly a 3x speedup.

The implementation here looks pretty exotic, but it's very inspired by
Go's own `encoding/json`.

Testing is pretty extensive, so I have pretty good confidence that this
works as expected based off of the passing suite.

Fixes part of #449.