Fix bug where Params.AppendTo not called if leaf had its own implementation
In Go you can "embed" a struct into another struct [1]:

    type ChargeParams struct {
        Params
    }

Embeds mostly behave like normal fields, except that they have the
special property that they allow the leaf struct to inherit any
functions implemented on the embedded struct.

For `Params`, this means that most parameter structs inherited the
`AppendTo` implementation of basic `Params`, which is where we encoded
our "extra" information.

However, when a leaf declares its own function by the same name, it no
longer inherits its embedded struct's implementation. This means that
any params that declare their own `AppendTo` will not currently encode
extras, because the `AppendTo` of `Params` doesn't get called.

This patch shifts things around a little bit so that we move encoding of
extras into an `AppendTo` on a new struct `ExtraValues`, which is an
alias for `url.Values`. This is breaking if a user was initializing
`Extra` directly, but hopefully most users were using the preferred
`params.AddExtra` instead. Regardless, this is the least invasive way I
could think of to fix this, so I think we'll have to go through with it.

We also do something similar for filters in `ListParams` (although no
`ListParams` implements an `AppendTo` so this is more of a hedge against
possible future problems) and change some tests around to better detect
these conditions (it wasn't very well tested before which is why the bug
happened).

[1] https://golang.org/doc/effective_go.html#embedding