Introduce extra concurrency safety for `form` package
This fixes #469 which was a problem where concurrent protection in the
caches wasn't working properly, making multiple closely-spaced
simultaneous accesses would result in stored maps throwing an error
because they were being read and written at the same time:

    fatal error: concurrent map read and map write

Here we introduce an `RWMutex` which allows unlimited concurrent reads,
but no reads if any single write operation is in progress (see [1] for
details). It also drops the complexity around using `atomic.Value` now
that we have every operation wrapped in a mutex anyway.

I added a simple test case that I found was able to reproduce the
problem ~100% of the time, and these changes see it pass now on every
attempt.

Fixes #469.

[1] https://godoc.org/sync#RWMutex