Skip file upload test for now
We're seeing some intermittent unreliability from this test case right
now (the owning team is looking into it), so let's put in a skip so that
we keep working on getting some of these other Go changesets in.

Note that we still compile the test, so at least we get a baseline
guarantee that things aren't too wildly incorrect.