Error early when params not specified for card functions
Card-related API calls are a bit of an unusual case because we need to
have an instantiated params struct in order to determine the correct URL
to which to make the request. Previously, if a `nil` was passed to any
of these functions, we'd panic as we tried to access one of its members.

This patch just makes the params requirement a little more explicit by
returning an error if any of card functions is called with `nil` params.

Fixes #483.