Rename language fields in Stripe's user agent to start with `lang`
I noticed this as I was looking at a dashboard that @ob-stripe produced
today. Other languages send language and language version as `lang` and
`lang_version` while we were using the fully expanded `language_*` here.
I'm changing it here for consistency with other libraries, but I'm also
going to patch the server because there are so many Go libraries out
there that could potentially be reporting if we supported both styles.