Handle all `json.Unmarshal` errors
We had a user over in #381 wondering why the errors from a
`json.Unmarshal` in `payout.go` wasn't being handled.

I believe the answer is that we've had a convention in here forever
whereby certain unmarshals from JSON blobs that are "low risk" are
assumed to succeed often enough that the error isn't worth handling. In
some of these `Unmarshal` implementations in particular, the `data` has
already been run through a previous `Unmarshal` so it's known that it
contains valid JSON and is unlikely to error on a second run.

Given that we've never had a complain about this before, it's probably
true that these `Unmarshals` are low risk and probably fine as is, but
nonetheless, it's still good practice to make sure that every error is
handled just in case. In this patch I've gone back through to any
`Unmarshal` invocations that were missing a check and added one.