Encode empty map values
Because most parameter fields in stripe-go are non-pointers, we
generally require that a value be non-empty before we encode it, because
otherwise we can't tell whether or not it was explicitly set.

It recently came up in #511 that it's no longer possible to set an empty
value in a map, which is a direct result of this behavior.

This patch tweaks the implementation slightly so that we always encode
all values that we find in maps, even if they're zero values. This makes
sense because unlike a struct which has default fields, we can always
guarantee that anything present in a map *was* explicitly added. This
resolves the problem described above.

Fixes #511.