Reverts protections against number of parameter mismatches
Reverts 78d772a34f04e25caed47586dcfd3bf37b5fde2e which added some
protections against mismatches in number of parameters.

This was an interesting experiment, but it adds some runtime overhead,
and so far this hasn't been a huge problem. I'm taking it back out but
leaving it in Git history in case we want to resurrect it at some point.