Activate linting for a few packages
This patch brings golint [1] into the Travis build and adds a make
target that will start running it on a few packages that we've expunged
of all linting errors.

I've just added a few packages to this initially because there are
hundreds of linting errors across the entire system, and cleaning them
all up would be an incredibly noisy diff. My recommendation is that we
issue smaller patches with ~1-6 packages converted over at a time to
make reviews easier (and more reliable), and add them to the whitelist
in `Makefile` as we go. After all packages are converted, we change the
call to `golint ./...`, and it'll always just run for everything.

Golint's major advancement will be to help us vet that Godoc comments
are formatted correctly, e.g. they all start with a function name like
`Get retrieves a charge`, and that exported functions _have_ Godoc
comments. This will hopefully get us to a place where we can generate
and publish a complete set of clean Godocs.

[1] https://github.com/golang/lint