Use `travis_retry` helper when using `go get` for dependencies
Our resolved dependencies aren't cached and we occasionally see
intermittent failures as `go get` runs and temporarily fails to get one
of our dependencies. This fails the build, and someone has to go in and
restart it.

It turns out that Travis provides a handy `travis_retry` helper [1]
which will retry a command up to three times if it returns a non-zero
exit code. I've added it in for all dependency resolution, and we'll see
if it helps.

[1] https://docs.travis-ci.com/user/common-build-problems/#travis_retry