Set request body before every retry
This patch changes the interfaces of `NewRequest` and `Do` around a
little so that we can set a new request body with every request.

In the era of HTTP (1), it was safe to reuse a `Request` object, but
with the addition of HTTP/2, it's now only sometimes safe. Reusing a
`Request` with a body will break.

See some more information here:

https://github.com/golang/go/issues/19653#issuecomment-341539160

Fixes https://github.com/stripe/stripe-go/issues/642.