All `BackendImplementation`s should sleep by default on retries
In #648 I introduced a new field on `BackendImplementation` which allows
us to disable sleeping between retries in tests. Unfortunately, I forgot
to set this field to `true` by default, and therefore we weren't
sleeping on retries in production.

Here we make sure to enable sleep when generating a new backend.

Fixes #651.