Remove the form package's `zero` option
Removes the form package's `zero` option, which was previously used to
encode a true zero for an integer value before we introduced pointers
everywhere.

I also looked into removing the `empty` option, but I think we still
need it for now. We could move it into a custom `AppendTo`
implementation specifically on `LegalEntityParams`, but it doesn't seem
like a clear win to me.