Remove racy test in favor of simpler one
Removes a test that's producing racy results in CI. As noted in #672, it
probably became racy as it started using the new HTTP implementation
in Go's HTTP/2 package.

My justification here is that cancellable contexts are already well
tested in Go's core. What we really want to test is just that the
context supplied by our users does indeed make it into the Go `Request`
that we're about to execute -- from there we can expect Go to do the
right thing. We already have a test that checks that just a few lines
above.