Remove `TestBankAccountGet_ByAccount` test
Unfortunately the test suite has a few tests that are broken when
running against the latest version of stripe-mock.

Two out of three of them are fixed after we bring in the OpenAPI
specification that exposes the `object` parameter in the customer
sources list endpoint and apply the stripe-mock patch which allows it to
correctly validate query parameters [1].

Unfortunately though, this last test still doesn't work because `object`
is not exposed on `/accounts/:id/external_accounts`. We could look into
enabling that, but given we at least have some working coverage for our
list invocation now, it's probably easier just to remove this test and
rely on the other one.

[1] https://github.com/stripe/stripe-mock/pull/108