Encode `Params` in `AppendToAsSourceOrExternalAccount`
`BankAccount` has a special encoding function that's invoked from its
`New` called `AppendToAsSourceOrExternalAccount`. It works quite a bit
differently from any other resource (with the exception of card) because
this special encoding is invoked then the function uses a `CallRaw`
instead of a `Call`, so no general `AppendTo` is ever used.

As described in #685, this caused a problem whereby the standard set of
fields in `Params` were not being included because the function didn't
have any consideration for them (although strangely, we *were* handling
them from card's equivalent). Here we add an allowance for `Params` and
a test to make sure that metadata is being encoded correctly.

Fixes #685.