Set `Request.GetBody`
Here we set `Request.GetBody`, which is used under certain circumstances
by the internal `net/http` core like in the case of a redirect. This
will usually not end up being used, but it doesn't hurt to set it in
case it is (it's run lazily, and few additional resources are needed in
case it's not run).

Fixes #710.