Add section in README on Go modules + simplify lead-in
Modifies the intro of the README to add explicit instructions on how to
use stripe-go with and without Go modules, including commands that need
to be run and what imports should look like.

I've also removed the "versioning" section because I don't think this
information is that all that novel, and certainly isn't useful enough to
occupy prime real estate at the front of the file.