Revert Go module support
Reverts:
https://github.com/stripe/stripe-go/pull/712

(And an earlier PR that added a `go.mod` file in the first place.)

Unfortunately, there are still some fairly bad incompatibilities between
Go modules and Dep for users still on the latter. For now, we're going
to revert Go module support because there doesn't seem to be any easy
way (i.e., anything that's not maintaining a fork) of doing a good job
of both. Projects can still use the Go module system, but will have to
pull this package in as a pre-module incompatible one.

We are hoping to eventually see basic module awareness merged into Dep:

https://github.com/golang/dep/pull/1963

If/when that gets done, we'll revert this revert after a small grace
period, and hopefully be back to a place where both packaging systems
are well supported.

I'll also modify the README to clarify the current situation in another
follow up PR.