Allow API arrays to be emptied by setting an empty array
Allows API arrays like `additional_owners` to be emptied by setting
their value to an explicit empty array. To accomplish this we need to
rework parts of the form-encoding library to detect this specific
condition.

I also had to rework a fair bit of the form package's test suite because
the existing code was making no distinction between an actual value
being set an empty in the generated form versus one that's not present.

Fixes #788.