Add test script to check that all clients are included in `client.API`
When adding a new package with a `Client` struct, it's really easy to
forget to add it to the big list of clients in `client.API`, and we
often have to rely on user reports to tell us that we forgot about it.

Here we add a script to run with CI which looks for packages with a
`Client` and compares that list against what's in `client.API`, then
errors if any are missing from the latter.

Also adds a few clients that were missing from the list.