Log only to info on 402 errors from Stripe
As discussed in #869, breaks apart how Stripe errors are logged so that
status 402 errors go to info instead of error. This is meant to signal
that those errors are probably prompted by user data (as opposed to
sending invalid parameters or something), and handling them differently
allows integrations to avoid filling their error logs with messages that
they can't really prevent.

Fixes #869.